requirements.txt
flask
pymongo
faiss-cpu
sentence-transformers
requests
beautifulsoup4
anthropic
python-dotenv
numpy